feat(capacity): upgrade VRAM tracking from u32 to u64
AtomicU32/u32 capped at ~4.3 PB; AtomicU64/u64 supports up to ~18.4 EB.
Updates VramPoolPort trait, VramPool, DistributedVramPool, analyzer,
placement_planner, and dashboard_handlers to use u64 throughout.

+++ crates/veronex/src/application/use_cases/placement_planner.rs
@@ -219,7 +219,7 @@ async fn planner_tick(
         .collect();
 
     // Provisional free VRAM tracking (prevents multi-model collision in same cycle)
-    let mut provisional_free: HashMap<Uuid, u32> = HashMap::new();
+    let mut provisional_free: HashMap<Uuid, u64> = HashMap::new();
     for p in &scale_out_candidates {
         provisional_free.insert(p.id, vram_pool.available_vram_mb(p.id));
     }
