+++ crates/veronex/src/infrastructure/outbound/capacity/vram_pool.rs
@@ -12,11 +12,11 @@ use crate::application::ports::outbound::concurrency_port::{
 ///
 /// Scope: model × provider pair. All `Atomic*` fields are concurrency-safe.
 struct ModelState {
-    weight_mb: u32,
+    weight_mb: u64,
     is_loaded: bool,
-    kv_per_request_mb: u32,
+    kv_per_request_mb: u64,
     /// Active KV cache reservations (in MB) for this model.
-    active_kv_mb: Arc<AtomicU32>,
+    active_kv_mb: Arc<AtomicU64>,
     /// Number of active requests (for reporting).
     active_count: Arc<AtomicU32>,
     /// Adaptive concurrency limit (0 = unlimited).
@@ -56,12 +56,12 @@ struct ModelState {
 }
 
 impl ModelState {
-    fn new(weight_mb: u32, is_loaded: bool, kv_per_request_mb: u32, max_concurrent: u32) -> Self {
+    fn new(weight_mb: u64, is_loaded: bool, kv_per_request_mb: u64, max_concurrent: u32) -> Self {
         Self {
             weight_mb,
             is_loaded,
             kv_per_request_mb,
-            active_kv_mb: Arc::new(AtomicU32::new(0)),
+            active_kv_mb: Arc::new(AtomicU64::new(0)),
             active_count: Arc::new(AtomicU32::new(0)),
             max_concurrent: AtomicU32::new(max_concurrent),
             baseline_tps: AtomicU32::new(0),
@@ -84,16 +84,16 @@ impl ModelState {
 
 /// Per-provider VRAM state.
 struct ProviderVramState {
-    total_mb: AtomicU32,
+    total_mb: AtomicU64,
     /// Global KV reservation counter across all models.
-    reserved_kv_mb: Arc<AtomicU32>,
+    reserved_kv_mb: Arc<AtomicU64>,
     /// Safety buffer (in permil, e.g. 200 = 20%). Increases on OOM.
     safety_permil: AtomicU32,
     /// Model name → model state.
     models: DashMap<String, ModelState>,
     /// Cached sum of weight_mb for all currently-loaded models (O(1) reads).
     /// Updated atomically on mark_model_loaded / mark_model_unloaded.
-    cached_loaded_weight_mb: AtomicU32,
+    cached_loaded_weight_mb: AtomicU64,
 
     // ── Phase 7 fields ──────────────────────────────────────────────────
 
@@ -146,11 +146,11 @@ impl VramPool {
             .entry(provider_id)
             .or_insert_with(|| {
                 Arc::new(ProviderVramState {
-                    total_mb: AtomicU32::new(0),
-                    reserved_kv_mb: Arc::new(AtomicU32::new(0)),
+                    total_mb: AtomicU64::new(0),
+                    reserved_kv_mb: Arc::new(AtomicU64::new(0)),
                     safety_permil: AtomicU32::new(DEFAULT_SAFETY_PERMIL),
                     models: DashMap::new(),
-                    cached_loaded_weight_mb: AtomicU32::new(0),
+                    cached_loaded_weight_mb: AtomicU64::new(0),
                     is_standby: AtomicBool::new(false),
                     transition_until: AtomicU64::new(0),
                     last_mem_available_mb: AtomicU32::new(0),
@@ -164,7 +164,7 @@ impl VramPool {
     /// Total weight of loaded models — O(1) via cached_loaded_weight_mb.
     /// Updated atomically by mark_model_loaded / mark_model_unloaded.
     #[inline]
-    fn loaded_weight_mb(state: &ProviderVramState) -> u32 {
+    fn loaded_weight_mb(state: &ProviderVramState) -> u64 {
         state.cached_loaded_weight_mb.load(Ordering::Acquire)
     }
 
@@ -231,6 +231,7 @@ impl VramPool {
         let safety = total * state.safety_permil.load(Ordering::Acquire) as i64 / 1000;
         total - loaded - kv - DEFAULT_BUFFER_MB as i64 - safety
     }
+
 }
 
 impl Default for VramPool {
@@ -363,21 +364,21 @@ impl VramPoolPort for VramPool {
         None
     }
 
-    fn total_vram_mb(&self, provider_id: Uuid) -> u32 {
+    fn total_vram_mb(&self, provider_id: Uuid) -> u64 {
         self.providers
             .get(&provider_id)
             .map(|s| s.total_mb.load(Ordering::Acquire))
             .unwrap_or(0)
     }
 
-    fn used_vram_mb(&self, provider_id: Uuid) -> u32 {
+    fn used_vram_mb(&self, provider_id: Uuid) -> u64 {
         self.providers
             .get(&provider_id)
             .map(|s| Self::loaded_weight_mb(&s) + s.reserved_kv_mb.load(Ordering::Acquire))
             .unwrap_or(0)
     }
 
-    fn available_vram_mb(&self, provider_id: Uuid) -> u32 {
+    fn available_vram_mb(&self, provider_id: Uuid) -> u64 {
         self.providers
             .get(&provider_id)
             .map(|s| {
@@ -385,23 +386,22 @@ impl VramPoolPort for VramPool {
                 if raw == i64::MAX {
                     // VRAM not probed (APU/iGPU unified memory).
                     // Return a concurrency-headroom-based score so unprobed providers
-                    // compete fairly with VRAM-probed providers instead of always winning
-                    // (i64::MAX as u32 overflows to u32::MAX = 4,294,967,295).
+                    // compete fairly with VRAM-probed providers instead of always winning.
                     let active = s.total_active_count.load(Ordering::Acquire);
                     let mc = s.models.iter()
                         .map(|e| e.max_concurrent.load(Ordering::Acquire))
                         .max()
                         .unwrap_or(4);
                     // 1,024 MB per free slot — at least 1 so it's not filtered out
-                    mc.saturating_sub(active).saturating_mul(1_024).max(1)
+                    mc.saturating_sub(active).saturating_mul(1_024).max(1) as u64
                 } else {
-                    raw.max(0) as u32
+                    raw.max(0) as u64
                 }
             })
             .unwrap_or(0)
     }
 
-    fn set_total_vram(&self, provider_id: Uuid, total_mb: u32) {
+    fn set_total_vram(&self, provider_id: Uuid, total_mb: u64) {
         let state = self.get_or_create(provider_id);
         state.total_mb.store(total_mb, Ordering::Release);
     }
@@ -418,7 +418,7 @@ impl VramPoolPort for VramPool {
             .or_insert_with(|| ModelState::new(profile.weight_mb, false, profile.kv_per_request_mb, 0));
     }
 
-    fn mark_model_loaded(&self, provider_id: Uuid, model: &str, weight_mb: u32) {
+    fn mark_model_loaded(&self, provider_id: Uuid, model: &str, weight_mb: u64) {
         let state = self.get_or_create(provider_id);
         // Update cached weight: subtract old weight if already loaded, add new weight.
         let prev = state.models.get(model).map(|ms| (ms.is_loaded, ms.weight_mb));
@@ -726,7 +726,7 @@ impl VramPoolPort for VramPool {
             .unwrap_or(0)
     }
 
-    fn model_weight_mb(&self, provider_id: Uuid, model: &str) -> u32 {
+    fn model_weight_mb(&self, provider_id: Uuid, model: &str) -> u64 {
         self.providers
             .get(&provider_id)
             .and_then(|s| s.models.get(model).map(|ms| ms.weight_mb))
@@ -872,13 +872,13 @@ mod tests {
 
     // ── compute_available tests ──────────────────────────────────────────
 
-    fn make_provider_state(total: u32, kv: u32, safety: u32) -> ProviderVramState {
+    fn make_provider_state(total: u64, kv: u64, safety: u32) -> ProviderVramState {
         ProviderVramState {
-            total_mb: AtomicU32::new(total),
-            reserved_kv_mb: Arc::new(AtomicU32::new(kv)),
+            total_mb: AtomicU64::new(total),
+            reserved_kv_mb: Arc::new(AtomicU64::new(kv)),
             safety_permil: AtomicU32::new(safety),
             models: DashMap::new(),
-            cached_loaded_weight_mb: AtomicU32::new(0),
+            cached_loaded_weight_mb: AtomicU64::new(0),
             is_standby: AtomicBool::new(false),
             transition_until: AtomicU64::new(0),
             last_mem_available_mb: AtomicU32::new(0),
